Extract the ttl related configuration stuff into and different file and fix and a potential bug
