+++ src/main/scala/akka/persistence/dynamodb/DynamoDBTTLConfig.scala
@@ -1,0 +1,40 @@
+package akka.persistence.dynamodb
+
+import com.typesafe.config.Config
+
+import java.time.OffsetDateTime
+import scala.concurrent.duration.{ Duration, DurationLong }
+import scala.util.Try
+
+case class DynamoDBTTLConfig(fieldName: String, ttl: DynamoDBTTL) {
+
+  override def toString: String =
+    s"DynamoDBTTLConfig(fieldName = ${fieldName}, ttl = ${ttl.ttl})"
+}
+
+case class DynamoDBTTL(ttl: Duration) {
+  def getItemExpiryTimeSeconds(now: OffsetDateTime): Long =
+    now.plusNanos(ttl.toNanos).toEpochSecond
+}
+
+object DynamoDBTTL {
+  def fromJavaDuration(duration: java.time.Duration): DynamoDBTTL =
+    DynamoDBTTL(duration.toNanos.nanos)
+}
+
+object DynamoDBTTLConfigReader {
+
+  val configFieldName: String = "dynamodb-item-ttl-config.field-name"
+  val configTtlName: String = "dynamodb-item-ttl-config.ttl"
+
+  def readTTLConfig(c: Config): Option[DynamoDBTTLConfig] = {
+    for {
+      fieldName <- Try(c getString configFieldName).toOption
+      if fieldName.trim.nonEmpty
+      ttl <- Try(c getDuration configTtlName).toOption
+    } yield DynamoDBTTLConfig(
+      fieldName = fieldName,
+      ttl       = DynamoDBTTL.fromJavaDuration(ttl))
+  }
+
+}
